Support chaining modes with external IV, and no internal padding
This is particularly useful when using the API in the chained mode,
where we don't call it encrypt/decrypt once for long
plaintext/ciphertext but rather call it repeatedly for block-sized
chunks.

Because some of the chaining modes modify IV, we need to output the
modified IV after one such encryption/decryption call, to be later used
for subsequent calls.

Additionally, BCRYPT_BLOCK_PADDING isn't useful since the user already
passes the block-sized chunks, which are instead padded to the next
block size when using this flag. This isn't desired and prevents from
using the API in "chained" mode, so allow the user to control the
behaviour via the new `symmetric::Padding` enum.