Fix unsafeties related to WindowsString and add tests
Previously, `WindowsString::from_ptr` was a safe function but
dereferenced a raw pointer value - that's memory unsafe.
Instead, a new `WindowsString::from_bytes_with_nul` is introduced that's
fallible, can reuse the passed allocation but is also much more explicit
about the format it expects the data in (NUL-terminated).

This also fixes the `ToString` implementation and, in general, makes the
NUL handling more explicit.