Improve ergonomics around symmetric key with compile-time known key sizes
This is required if someone is to use the BlockCipher trait
implementations as by default providing KeySize of U16 for Aes means
that it can be only used for key sizes of 128 bits in generic contexts.

Instead of that, allow for optionally specifying key sizes for key
handes at compile-time which can be used in both run and compile time
contexts.